newslens/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
httpx>=0.28.1

# Endee Python SDK
endee>=0.1.6

# Embedding model
sentence-transformers>=3.0.1
torch>=2.9.0
transformers>=4.41.0

# Data handling
pandas>=2.2.2
numpy>=2.0.0